+++ bin/packages/tar_package.js
@@ -1,15 +1,18 @@
 const FileSystem = require('../support/file_system')
+const StatusLog = require('../support/status_log')
 const tar = require('tar')
 
 class TarPackage {
   static pack (archiveBundle, manifestConfiguration) {
     const releaseAsset = `${archiveBundle.releaseStaging}.tar`
 
+    StatusLog.inform('pack', 'tar', { path: releaseAsset })
     return tar.create({
       gzip: false,
       file: FileSystem.readPath(releaseAsset)
     }, [archiveBundle.releaseFolder])
       .then(() => {
+        StatusLog.inform('packed', 'tar', { path: releaseAsset })
         return {
           archiveBundle: {
             releaseAsset
@@ -21,6 +24,7 @@ class TarPackage {
     cachePath,
     stagingPath
   }) {
+    StatusLog.inform('unpack', 'tar', { path: cachePath })
     return tar.extract({
       cwd: stagingPath,
       file: FileSystem.readPath(cachePath),
@@ -29,6 +33,7 @@ class TarPackage {
       strip: 1
     })
       .then(() => {
+        StatusLog.inform('unpacked', 'tar', { path: cachePath })
         return {}
       })
   }