+++ bin/transits/local_transit.js
@@ -1,3 +1,4 @@
+const StatusLog = require('../support/status_log')
 const fs = require('fs-extra')
 const _ = require('lodash')
 
@@ -9,6 +10,7 @@ class LocalTransit {
     const writeTo = this.__cleanURI(uri)
 
     return new Promise((resolve, reject) => {
+      StatusLog.inform('push', 'local', { from: releaseStaging, to: writeTo })
       fs.copy(releaseStaging, writeTo, err => {
         if (err) {
           reject(new Error(err))
@@ -23,6 +25,7 @@ class LocalTransit {
     const originalLocation = this.__cleanURI(uri)
 
     return new Promise((resolve, reject) => {
+      StatusLog.inform('pull', 'local', { from: originalLocation, to: cachePath })
       fs.copy(originalLocation, cachePath, err => {
         if (err) {
           reject(new Error(err))